# Improve abs url replacement speed