--- a/.gitignore
+++ b/.gitignore
@@ -2,6 +2,8 @@ hugo
 docs/public*
 hugo.exe
 *.test
+*.prof
+nohup.out
 cover.out
 *.swp
 *.swo