--- a/transform/post.go
+++ b/transform/post.go
@@ -1,1 +1,0 @@
-package transform