@spf13 2015-02-16
Great stuff. Merged as d33a42f